+++ lib/bs_flutter_buttons.dart
@@ -11,5 +11,7 @@ export 'src/bs_dropdown_button.dart';
 export 'src/widgets/bs_dropdown_menu.dart';
 export 'src/widgets/bs_dropdown_item.dart';
 export 'src/widgets/bs_dropdown_divider.dart';
+export 'src/widgets/bs_dropdown_header.dart';
 export 'src/config/bs_dropdown_menu_style.dart';
+export 'src/config/bs_dropdown_item_style.dart';
 export 'src/config/bs_dropdown_menu_size.dart';