+++ lib/src/config/bs_dropdown_item_style.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class BsDropdownItemStyle {
+
+  const BsDropdownItemStyle({
+    this.backgroundColor,
+    this.activeBackgroundColor = Colors.blueAccent,
+    this.disabledBackgroundColor,
+  });
+
+  final Color? backgroundColor;
+
+  final Color activeBackgroundColor;
+
+  final Color? disabledBackgroundColor;
+}